split even more
